little things
git-svn-id: svn://svn.powerdns.com/pdns/trunk/pdns@157 d19b8d6e-7fed-0310-83ef-9ca221ded41b

+++ pdns/resolver.cc
@@ -303,7 +303,7 @@ int Resolver::axfr(const string &ip, const char *domain)
   p.setQuestion(Opcode::Query,domain,QType::AXFR);
   p.wrapup();
 
-  int replen=htons(p.len);
+  u_int16_t replen=htons(p.len);
   Utility::iovec iov[2];
   iov[0].iov_base=(char*)&replen;
   iov[0].iov_len=2;
